+++ packages/react/src/components/DataTable/src/components/DataTableDropdown/index.ts
@@ -1,0 +1,7 @@
+export { DataTableDropdown } from "./DataTableDropdown";
+export {
+  DataTableDropdownContext,
+  useDataTableDropdownContext,
+} from "./DataTableDropdownContext";
+export type { DataTableDropdownProps } from "./dataTableDropdown.types";
+export type { DataTableDropdownContextValue } from "./dataTableDropdownContext.types";