+++ packages/react/src/components/DataTable/src/components/DataTableDropdownDivider/DataTableDropdownDivider.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import { Divider } from "@nimbus-ds/components";
+
+/**
+ * Divider component for separating groups of actions in the DataTable Dropdown.
+ */
+const DataTableDropdownDivider: React.FC = () => <Divider thickness={1} />;
+
+DataTableDropdownDivider.displayName = "DataTable.DropdownDivider";
+
+export { DataTableDropdownDivider };